feat: public courses (#1617)

+++ api/src/catalog/catalog.utils.ts
@@ -5,7 +5,11 @@
 import fs from 'fs/promises';
 import { request } from 'graphql-request';
 
-import { catalogBySeasonQuery, listSeasonsQuery } from './catalog.queries';
+import {
+  catalogBySeasonQuery,
+  catalogBySeasonNoRatingsQuery,
+  listSeasonsQuery,
+} from './catalog.queries';
 import { GRAPHQL_ENDPOINT, STATIC_FILE_DIR } from '../config';
 import winston from '../logging/winston';
 
@@ -90,41 +94,74 @@ export async function fetchCatalog(
 
   // For each season, fetch all courses inside it and save
   // (if overwrite = true or if file does not exist)
+
   const processSeasons = seasons.seasons.map(async (season) => {
     const seasonCode = season.season_code;
-    const outputPath = `${STATIC_FILE_DIR}/catalogs/${seasonCode}.json`;
+    const fullCatalogPath = `${STATIC_FILE_DIR}/catalogs/${seasonCode}.json`;
+    const publicCatalogPath = `${STATIC_FILE_DIR}/catalogs/public/${seasonCode}.json`;
+    let catalogFull: Catalog = { computed_listing_info: [] };
+    let catalogPublic: Catalog = { computed_listing_info: [] };
 
+    // Fetch and save the full catalog (including evaluations)
     if (
-      !overwrite &&
-      (await fs.access(outputPath).then(
+      overwrite ||
+      !(await fs.access(fullCatalogPath).then(
         () => true,
         () => false,
       ))
     ) {
-      winston.info(`Catalog for ${seasonCode} exists, skipping`);
-      return;
+      try {
+        catalogFull = await request(GRAPHQL_ENDPOINT, catalogBySeasonQuery, {
+          season: seasonCode,
+        });
+        await fs.writeFile(
+          fullCatalogPath,
+          JSON.stringify(catalogFull.computed_listing_info),
+        );
+        winston.info(
+          `Fetched full catalog for season ${seasonCode}: n=${catalogFull.computed_listing_info.length}`,
+        );
+      } catch (err) {
+        winston.error(
+          `Error fetching full catalog for season ${seasonCode}: ${err}`,
+        );
+      }
+    } else {
+      winston.info(`Full catalog for ${seasonCode} exists, skipping`);
     }
 
-    let catalog: Catalog = { computed_listing_info: [] };
-
-    try {
-      catalog = await request(GRAPHQL_ENDPOINT, catalogBySeasonQuery, {
-        season: seasonCode,
-      });
-    } catch (err) {
-      winston.error(err);
-      throw err;
+    // Fetch and save the public catalog (no evaluations)
+    if (
+      overwrite ||
+      !(await fs.access(publicCatalogPath).then(
+        () => true,
+        () => false,
+      ))
+    ) {
+      try {
+        catalogPublic = await request(
+          GRAPHQL_ENDPOINT,
+          catalogBySeasonNoRatingsQuery,
+          { season: seasonCode },
+        );
+        await fs.mkdir(`${STATIC_FILE_DIR}/catalogs/public`, {
+          recursive: true,
+        });
+        await fs.writeFile(
+          publicCatalogPath,
+          JSON.stringify(catalogPublic.computed_listing_info),
+        );
+        winston.info(
+          `Fetched public catalog for season ${seasonCode}: n=${catalogPublic.computed_listing_info.length}`,
+        );
+      } catch (err) {
+        winston.error(
+          `Error fetching public catalog for season ${seasonCode}: ${err}`, // MAKE SURE ITS MADE
+        );
+      }
+    } else {
+      winston.info(`Public catalog for ${seasonCode} exists, skipping`);
     }
-
-    await fs.writeFile(
-      outputPath,
-      JSON.stringify(catalog.computed_listing_info),
-    );
-
-    winston.info(
-      `Fetched season ${seasonCode}: n=${catalog.computed_listing_info.length}`,
-    );
   });
-
   return Promise.allSettled(processSeasons);
 }
